refactor: remove src package and update imports
